pyproject.toml
[project]
name = "flowmind-ai"
version = "0.1.0"
description = "AI adaptive traffic signal system for Vietnamese smart cities"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "eclipse-sumo>=1.20.0",
    "traci>=1.20.0",
    "sumolib>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
